Restructured the code to be more readable
