+++ server_cpp/src/strategy_io/IInputReader.h
