+++ server_cpp/src/strategy_io/console/ConsoleOutputWriter.cpp
@@ -2,7 +2,6 @@
 #include "ConsoleOutputWriter.h"
 #include <iostream>
 
-/// TODO add tests
 
 /**
 * Default constructor
