+++ server_cpp/src/strategy_io/tcp/TCPInputReader.h
@@ -1,8 +1,7 @@
 // Author(s): Yuval Anteby
 #ifndef TCP_INPUT_READER_H
 #define TCP_INPUT_READER_H
-/// TODO tests
-#include "IInputReader.h"
+#include "./strategy_io/IInputReader.h"
 
 /**
  * Class responsible on reading input from a user using a TCP connection to a client.
